backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.18.1
google-adk>=0.1.0
